backend/requirements.txt
django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
flake8==7.1.1
black==24.8.0
psycopg2-binary==2.9.9
requests==2.32.3
debugpy==1.8.5
